only permit configuration through env vars
This also adds an `example.env` file to explain
what options are available for configuration.
