+++ crates/astria-composer/example.env
@@ -1,0 +1,14 @@
+# Log level. One of debug, info, warn, or error
+ASTRIA_COMPOSER_LOG="astria_composer=info"
+
+# Address of the API server
+ASTRIA_COMPOSER_API_LISTEN_ADDR="0.0.0.0:0"
+
+# Address of the RPC server for the sequencer chain
+ASTRIA_COMPOSER_SEQUENCER_URL="http://127.0.0.1:26657"
+
+# Chain ID that we want to connect to
+ASTRIA_COMPOSER_CHAIN_ID="912559"
+
+# Address of the RPC server for execution
+ASTRIA_COMPOSER_EXECUTION_URL="ws://127.0.0.1:8545"
